cleanup: also clean hosts file on Windows
Hosts file on Windows also contains crc entries. crc cleanup should
remove them.
