+++ pkg/crc/preflight/preflight_linux_test.go
@@ -63,11 +63,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: true,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -96,11 +96,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -128,11 +128,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -154,11 +154,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: true,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -187,11 +187,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -219,11 +219,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -245,11 +245,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: true,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -278,11 +278,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -310,11 +310,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -336,11 +336,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: true,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -370,11 +370,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
@@ -403,11 +403,11 @@ var checkListForDistros = []checkListForDistro{
 		systemdResolved: false,
 		checks: []Check{
 			{check: checkIfRunningAsNormalUser},
-			{cleanup: removeHostsFileEntry},
 			{check: checkPodmanExecutableCached},
 			{check: checkAdminHelperExecutableCached},
 			{configKeySuffix: "check-ram"},
 			{cleanup: removeCRCMachinesDir},
+			{cleanup: removeHostsFileEntry},
 			{cleanup: removeOldLogs},
 			{cleanup: cluster.ForgetPullSecret},
 			{check: checkVirtualizationEnabled},
